+++ python/paddle/__init__.py
@@ -380,6 +380,7 @@ from . import hub  # noqa: F401
 from . import linalg  # noqa: F401
 from . import fft  # noqa: F401
 from . import signal  # noqa: F401
+from . import parallel  # noqa: F401
 
 import paddle.text  # noqa: F401
 import paddle.vision  # noqa: F401
